Вносит изменения
