+++ js/effects.js
@@ -1,55 +1,4 @@
-const EFFECTS = [
-  {
-    name: 'none',
-    style: 'none',
-    min: 0,
-    max: 100,
-    step: 1,
-    unit: '',
-  },
-  {
-    name: 'chrome',
-    style: 'grayscale',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'sepia',
-    style: 'sepia',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'marvin',
-    style: 'invert',
-    min: 0,
-    max: 100,
-    step: 1,
-    unit: '%',
-  },
-  {
-    name: 'phobos',
-    style: 'blur',
-    min: 0,
-    max: 3,
-    step: 0.1,
-    unit: 'px',
-  },
-  {
-    name: 'heat',
-    style: 'brightness',
-    min: 1,
-    max: 3,
-    step: 0.1,
-    unit: '',
-  }
-];
-
-const DEFAULT_EFFECT = EFFECTS[0];
+import { EFFECTS } from './constants.js';
 
 const previewElement = document.querySelector('.img-upload__preview img');
 const sliderContainerElement = document.querySelector('.img-upload__effect-level');
@@ -57,9 +6,10 @@ const sliderElement = document.querySelector('.effect-level__slider');
 const effectValueElement = document.querySelector('.effect-level__value');
 const pictureEffectsElement = document.querySelector('.img-upload__effects');
 
-let currentEffect = DEFAULT_EFFECT;
+const defaultEffect = EFFECTS[0];
+let currentEffect = defaultEffect;
 
-const isDefault = () => currentEffect === DEFAULT_EFFECT;
+const isDefault = () => currentEffect === defaultEffect;
 
 const onShowSlider = () => {
   sliderContainerElement.classList.remove('hidden');
@@ -102,18 +52,18 @@ pictureEffectsElement.addEventListener('change', (evt) => {
 const onSliderUpdate = () => {
   const sliderValue = sliderElement.noUiSlider.get();
   previewElement.style.filter = isDefault()
-    ? DEFAULT_EFFECT.style
+    ? defaultEffect.style
     : `${currentEffect.style}(${sliderValue}${currentEffect.unit})`;
   effectValueElement.value = sliderValue;
 };
 
 noUiSlider.create(sliderElement, {
   range: {
-    min: DEFAULT_EFFECT.min,
-    max: DEFAULT_EFFECT.max,
+    min: defaultEffect.min,
+    max: defaultEffect.max,
   },
-  start: DEFAULT_EFFECT.max,
-  step: DEFAULT_EFFECT.step,
+  start: defaultEffect.max,
+  step: defaultEffect.step,
   connect: 'lower',
 });
 onCloseSlider();
@@ -123,8 +73,8 @@ sliderElement.noUiSlider.on('update', () => {
 });
 
 const resetEffects = () => {
-  currentEffect = DEFAULT_EFFECT;
+  currentEffect = defaultEffect;
   updateSlider();
 };
 
-export {resetEffects};
+export { resetEffects };
