+++ js/scale.js
@@ -1,10 +1,11 @@
-import { GetRange } from './util.js';
+import { Range } from './util.js';
 
 const previewElement = document.querySelector('.img-upload__preview img');
 const scaleValueElement = document.querySelector('.scale__control--value');
 const scaleIncreaseElement = document.querySelector('.scale__control--bigger');
 const scaleDecreaseElement = document.querySelector('.scale__control--smaller');
-let scale = new GetRange(25, 100, 25, 100);
+
+let scale = new Range(25, 100, 25, 100);
 
 const updateScale = () => {
   scaleValueElement.value = `${scale.value}%`;
@@ -13,7 +14,7 @@ const updateScale = () => {
 };
 
 const resetScale = () => {
-  scale = new GetRange(25, 100, 25, 100);
+  scale = new Range(25, 100, 25, 100);
   updateScale();
 };
 
@@ -32,4 +33,5 @@ const clickScaler = () => {
   scaleDecreaseElement.addEventListener('click', onScaleDecreaseClick);
   updateScale();
 };
+
 export { resetScale, clickScaler };
