+++ js/util.js
@@ -1,10 +1,12 @@
-const COMMENT_STEP_COUNT = 5;
+import { COMMENT_STEP_COUNT } from './constants.js';
 
 const getInitialCommentStateCount = (count = 0) => {
   let countInitialCommentState = count;
   countInitialCommentState += COMMENT_STEP_COUNT;
   return countInitialCommentState;
 };
+
+const isEscapeKey = (evt) => evt.key === 'Escape';
 
 const removeEventListener = (removeEventListenerWhere, typeOfEvent, onEventFunction) => {
   const element = removeEventListenerWhere;
@@ -29,29 +31,27 @@ const getRandomInteger = (a, b) => {
 const getRandomArrayElement = (array) =>
   array[getRandomInteger(0, array.length - 1)];
 
-const isEscapeKey = (evt) => evt.key === 'Escape';
-
-function GetRange(min = 0, max = 100, step = 1, value) {
+const Range = function(min = 0, max = 100, step = 1, value) {
   this._min = min;
   this._max = max;
   this._step = step;
   this.value = value || max;
-}
+};
 
-GetRange.prototype.increase = function() {
+Range.prototype.increase = function() {
   this.value = Math.min(this.value + this._step, this._max);
 };
 
-GetRange.prototype.decrease = function() {
+Range.prototype.decrease = function() {
   this.value = Math.max(this.value - this._step, this._min);
 };
 
 export {
-  COMMENT_STEP_COUNT,
-  getInitialCommentStateCount,
-  removeEventListener,
   generateUniqueNumber,
   getRandomInteger,
   getRandomArrayElement,
   isEscapeKey,
-  GetRange };
+  removeEventListener,
+  getInitialCommentStateCount,
+  Range
+};
