+++ js/validation.js
@@ -58,3 +58,5 @@ const setFormSubmit = () => {
 };
 
 export {setFormSubmit, pristine, unblockSubmitButton};
+
+
